requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.7.0
fastapi==0.109.2
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.3.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
motor==3.3.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.1
pydantic-settings==2.2.1
pydantic_core==2.33.0
pymongo==4.6.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
supervision==0.24.0
sympy==1.13.1
thop==0.1.1.post2209072238
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.1.27
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.27.1
websockets==12.0
Werkzeug==3.1.3
